requirements.txt
anthropic>=0.40.0
openai>=1.50.0
sentence-transformers>=3.0.0
qdrant-client>=1.12.0
langchain>=0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
redis>=5.0.0
rank-bm25>=0.2.2
pdfplumber>=0.11.0
psycopg2-binary>=2.9.0
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
locust>=2.30.0
